server/requirements.txt
gunicorn
requests
flask
Pillow
numpy
